+++ src/cjs/utils/algorithms/matrixFill.js
@@ -13,7 +13,7 @@ var arrayFill_1 = require("./arrayFill");
 var matrixFill = function (countA, countB, initialValue) {
     var arr = Array(countA);
     for (var i = 0; i < countA; i++) {
-        arr[i] = (0, arrayFill_1.arrayFill)(countB, initialValue);
+        arr[i] = arrayFill_1.arrayFill(countB, initialValue);
     }
     return arr;
 };